# [bugfix-1.1.x] COREXY stutter moves (planner.cpp changes) 