This is from Issue #8573 

In a band of feed rates the CORE machines will have short (10mS – 30mS) pauses.

The root cause is the current code doesn’t apply jerk until the head movement rate exceeds the jerk setting.  On a CORE machine one of the steppers is always moving faster than the head.  The result is a band of feed rates where the stepper needs jerk but the head rate isn’t yet calling for it.

In that band of feed rates, negative values for **block->accelerate_until** are fed to the stepper routine.  These negative values result in the short pauses of about 10mS - 30mS spaced about 275mS apart..
 
This update calculates the jerk requirements on a per stepper basis.  It also applies the worst case rather than just the one with the maximum configured jerk.

`nominal_speed` is now based on maximum stepper speed and `start_speed` is based on the worst case jerk.

----

CPU loading is not affected by the code change.  Execution times for `_buffer_steps()` are similar.

Execution times are in milliseconds.  

```
current  1.499  2.008  2.258  2.326  2.212  2.143
         1.533  1.977  2.258  2.326  2.212  2.125
         1.546  1.973  2.258  2.319  2.392  2.010
         1.505  2.002  2.211  2.380  2.206  2.136

update   1.616  1.900  2.271  2.321  2.254  2.126
         1.627  1.905  2.271  2.321  2.312  2.033
         1.627  1.901  2.251  2.321  2.102  2.207
         1.632  1.905  2.288  2.321  2.324  2.033
```

The above uses the following gcode program
```
  g1 x100 y100 f1000
  g1 x50  y200 f1000
  g1 x0   y0   f1000
  g1 x275 y25  f1000
  g1 x100 y100 f100
```

----

Build sizes and RAM usage are the same or a little less.
```
         text       data     bss 
current  51642       192    2406
update   51776       192    2406  
```

----

Here's an example of the stuttering when the feed rate is in problem band.
![image](https://user-images.githubusercontent.com/20692583/33699068-6ef8f7f6-dad6-11e7-9c58-d7da5c2bbec9.png)

Here are logic analyzer captures of the step pulses.  They include new & old software, COREXY & Cartesian and various feedrates.  The gcode program is given above.  X jerk was set to 20 and Y was set to 10 so that it is possible to see what happens when axis need different accelerations.

The very bottom trace is the time spend in the `_buffer_steps()` routine.

You'll need the [Saleae software](https://www.saleae.com/downloads) to view them.

[stepper captures.zip](https://github.com/MarlinFirmware/Marlin/files/1537848/stepper.captures.zip)