@thinkyhead 2017-12-07
This is interesting, and kind of scary… because it looks like it throws away a bunch of carefully-crafted jerk code, and kind of seems like a regression.

Has some other firmware taken this approach, and are we sure it's safe compared to all the other possible approaches that could be taken? After all, we've been flirting with possibly using an even fancier jerk method that takes angular change into account… and this PR goes the opposite way, towards what looks like utter simplicity.

I want to believe… and merge… but it seems too easy.