@Bob-the-Kuhn 2017-12-07
WOW!! That was merged a lot faster than I expected.

I'm also nervous about going away from the old method.  

In a simple Cartesian system the amount of energy supplied by the X stepper to make a change in the X direction of the head does not depend on what is happening in the Y or Z directions.

I'll need to do some thought experiments before I'm confident in how a CORE system works.  

Delta math is still a mystery to me.