@thinkyhead 2017-12-08
If the new Jerk code ends up in 1.1.7 (coming soon!) then we'll need to give it a fast sounding nickname in the release notes. The "Rocky Rakuhn" Jerk Algorithm…