@thinkyhead 2017-12-08
> In a simple Cartesian system the amount of energy supplied by the X stepper to make a change in the X direction of the head does not depend on what is happening in the Y or Z directions.

You're down the rabbit hole now! As you can see the core underlying system of Marlin presumes Cartesian mechanics to a fault. Delta was bolted on by @jcrocholl for the Kossel, and so there is no formal system in place to deal with the linear acceleration of the effector in a Delta or with the rotational degrees-per-second feedrates of a SCARA. So, we continue to apply patches and look for patterns in the chaos. But for the most part, each problem we tackle ends up having scattered pieces.

That said, the rates of motion in a Core system —in my view— should be considered in terms of the carriages, with the feedrates reckoned and limited at the `X_HEAD`/`Y_HEAD` level, and there should be no limits applied to the AB level. Their speeds should be whatever the result is when limits are applied in XY terms.

So, maybe that means there's a part of the stepper ISR that could be skipped or replaced, if its job is to apply unnecessary limits to AB that it should be applying to XY. Or, maybe that's what it already tries to do…?